Change addCourse in server to correct file, change route, and other minor fixes
