+++ server/src/controllers/course.ts
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: MIT
 
+import { Request, Response } from 'express';
+
 export interface LocalizedString {
   fi: string,
   sv: string,
@@ -16,4 +18,19 @@ export interface Course {
   department: LocalizedString,
   name: LocalizedString,
   evaluationInformation: LocalizedString
+}
+
+export async function addCourse(req: Request, res: Response): Promise<void> {
+  try {
+    // TODO: add the course to the database
+    res.send({
+      success: true
+    });
+  } catch (error) {
+    res.status(401);
+    res.send({
+      success: false,
+      error: error,
+    });
+  }
 }
