+++ server/src/controllers/user.ts
@@ -99,19 +99,3 @@ export async function getUserCourses(req: Request, res: Response): Promise<void>
     });
   }
 }
-
-export async function addCourse(req: Request, res: Response): Promise<void> {
-  try {
-    // TODO: add the course to the database
-    res.send({
-      success: true,
-      course: req.body,
-    });
-  } catch (error) {
-    res.status(401);
-    res.send({
-      success: false,
-      error: error,
-    });
-  }
-}
