+++ server/src/routes/index.ts
@@ -3,14 +3,15 @@
 // SPDX-License-Identifier: MIT
 
 import express, { Request, Response, Router } from 'express';
-import { getUserCourses, addCourse } from '../controllers/user';
+import { getUserCourses } from '../controllers/user';
+import { addCourse } from '../controllers/course';
 import { testDbFindAllUsers, testDbFindAllCourses } from '../controllers/test';
 
 export const router: Router = Router();
 
 router.get('/user/:userId/courses', getUserCourses);
 
-router.post('/user/:userId/courses', express.json(), addCourse);
+router.post('/v1/courses', express.json(), addCourse);
 
 // TODO: remove this test endpoint after working endpoint has been added
 router.get('/v1/test/db', testDbFindAllUsers);
